requirements.txt
alpaca-py>=0.30.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
pytest>=8.0.0
pytz>=2024.1
